Initial pipe-based output adapter.
It doesn't work, because the pipe buffers are too big!
The "random generator" program apparently leaks piles of
memory, but it's really just generating frames much faster
than they can be consumed. And it does not block before
it runs out of memory. Going to rewrite the pipe so that
this cannot happen so easily.