Initial work on the keying daemon
The keying daemon is made up of a keyer and several character generators.
The keyer is the main loop of the program; it retrieves overlays from the
character generators and mixes them into a video stream.

The character generators are responsible for generating the alpha overlay
images themselves. Each CG runs in its own thread and sends its output
as RawFrames through a Pipe to the keyer loop.

One CG class has been implemented: SvgSubprocessCharacterGenerator. Character
generators of this class communicate with an external script, possibly written
in some high-level language, to obtain overlay image data in SVG
(Scalable Vector Graphics) format. The SVG data is rendered using librsvg
and cairo and passed to the keyer. The SVG image is updated at each frame.