Built framework for scoreboard animation and control
Animation classes are defined in scoreboard/animate.rb.
Network control is via a TCP socket. This socket accepts commands
and parameters and routes them to the correct methods in the
scoreboard state class. A simple dissolve-in/out animation
has been developed for the RPI TV graphical theme.

Animations should probably be factored separately from the
scoreboard state. This will allow the theme to be changed
without hacking up the scoreboard code itself.