Make Ruby keyer wrappers work
See test.rb for an example of how this works. The keyer application is now
defined as a Ruby script that defines the C++ objects that do the actual
heavy lifting. Cool. Only problem is that the extconf.rb generates a bogus
Makefile that's missing a space. That is not fixed, so the Makefile must be
edited by hand.