avpinput_decklink_mjpeg: increase buffer size
The M-JPEG encoder uses a memory buffer to store the JPEG data
it creates. We make this buffer 16MB to completely eliminate any
possibility that the generated JPEG is larger than the buffer's
capacity, because this condition causes an exception to be raised
and the frame not to be encoded. Since this exception is unhandled,
the program also crashes if the buffer size is exceeded.