AudioPacket/AudioFIFO refactoring to DRY things up
AudioPacket has been split into PackedAudioPacket and PlanarAudioPacket.
These are templates e.g. PackedAudioPacket<int16_t>, PlanarAudioPacket<float>.
They can be converted via the make_planar<U>, make_packed<U>, and copy<U>
members.

AudioFIFO is also now a template and is used in most places where a fifo
is needed. The implementation should probably be converted into a circular
buffer. NOTE: AudioFIFO IS NOT THREAD SAFE! It is meant to be used to convert
a stream of packets into a stream of differently sized packets, within a
single thread.