keyer: add JsCharacterGenerator class
This character generator executes a piece of JavaScript code to render each
frame. Frames are built up from a collection of assets loaded from image
files. JavaScript is passed in from a subprocess via stdin and compiled
using v8 to yield an object; this object's "render" property is taken
as a rendering function and called for each frame. It calls "draw" to
combine the assets into the final frame.